// src/alterdata/admissoes/admissao.types.ts
export interface AdmissaoAttributes {
  nome: string;
  cpf: number;
  cpfAlfanumerico?: string;
  isCpf?: boolean;
  admissao: string;
  nomecargo: string;
  nascimento?: string;
  nomedamae?: string;
  nomedopai?: string;
  municipionascimento?: string;
  possuideficiencia?: boolean;
  observacao?: string;
  ocorrencia?: string;
  cep?: string;
  rua?: string;
  numero?: string | number;
  bairro?: string;
  complemento?: string;
  cidade?: string;
  email?: string;
  telefone?: string;
  celular?: string;
  dataatestadoocupacional?: string;
  diascontratoexperiencia?: number;
  dataterminocontrato?: string;
  primeiroemprego?: boolean;
  salario?: number;
  proventosedescontos?: string;
  requersegurodesemprego?: boolean;
  pis?: string;
  datapis?: string;
  cnh?: string;
  emissaocnh?: string;
  orgaoemissorcnh?: string;
  validadecnh?: string;
  primeiraemissaocnh?: string;
  orgaoemissoridentidade?: string;
  identidade?: string;
  dataidentidade?: string;
  ctps?: number;
  seriectps?: string;
  datactps?: string;
  tituloeleitor?: number;
  zonatituloeleitor?: string | number;
  secaotituloeleitor?: string | number;
  agencia?: string;
  conta?: string;
  historicosolicitacao?: string;
  banco?: string;
  valestrasnportes?: string;
  usuariocriacao?: string;
  datacriacao?: string;
  [key: string]: unknown;
}